% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeseries.R
\name{create_timeseries}
\alias{create_timeseries}
\title{create_timeseries}
\usage{
create_timeseries(
  email,
  fra,
  til,
  period = "SECOND",
  amount = 60,
  type = "contentId",
  contentId
)
}
\arguments{
\item{email}{Epost til konto du ønsker å spørre etter dataene med}

\item{fra}{Date(time) for starttidspunkt}

\item{til}{Date(time) for sluttidspunkt}

\item{period}{Nivået på periodene du ønsker å aggregere på. Tillater MICROSECOND , MILLISECOND , SECOND , MINUTE , HOUR eller DAY}

\item{amount}{Angi størrelsen på intervallene i perioden. Hvis period er satt til SECOND og amount til 20, vil vi lage et datasett med 20 sekunds intervaller}

\item{type}{Angi hva du ønsker å bruke som identifikator - contentId eller url.}

\item{ID}{Angi contentID for pluggen du ønsker å hente data for. ContentID behøver polypoly prefix'er og URL'er trenger å bli enkapsulert med %}
}
\value{
Et datasett som viser etasje og rom for en gitt artikkel på bestemte tidspunkter
}
\description{
Denne funksjonen lager et datasett med plassering av en plugg fra forsiden for en konkret periode med gitte tidsintervaller
}
\examples{
create_timeseries("eirik.brautaset(at)nrk.no","2022-04-03 00:00:00","2022-04-03 23:59:59",1,"SECONDS","pp:1.15918927")
create_timeseries("eirik.brautaset(at)nrk.no","2022-04-03 00:00:00","2022-04-03 23:59:59",1,"SECONDS",type="url", "\%paskenotter\%")

}
