% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCutoff.R
\name{get_cutoff}
\alias{get_cutoff}
\title{Get cut-off}
\usage{
get_cutoff(df, start, slutt)
}
\arguments{
\item{df}{Datasettet du ønsker evaluert. Her kan du benytte funksjonen get_df()}

\item{start}{starttidspunktet for programmets sendetid}

\item{slutt}{sluttidspunktet for programmets sendetid}
}
\value{
Et nytt datasett hvor all data i datasettet bqdata beregner ny secondsActive for spesifisert tidsintervall
}
\description{
Denne funksjonen tar imot et datasett med Kuratordata og beregner ny secondsVisible basert på et program sin sendetid
Funksjonen aksepterer p.t. bare hele timer
}
\examples{
cutOffDF = get_cutoff(bqdata,17,23)
cutOffDF = get_cutoff(get_df(sql,email),17,23)
}
