% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ.R
\name{get_df}
\alias{get_df}
\title{get_df}
\usage{
get_df(sql, epost)
}
\arguments{
\item{sql}{legg inn spørringen din her}

\item{email}{Epost til konto du ønsker å spørre etter dataene med}
}
\value{
Datasett fra BigQuery basert på spørringen du gjør i parameteret 'sql'
}
\description{
Forenklet funksjon for å hente data fra BigQuery
}
\examples{
get_df("SELECT * FROM 'nrk-datahub.consumer_facing_views.article_demography_total LIMIT' 1000","eirik.brautaset(at)nrk.no")

}
